ref(interaction container) creating interaction template component; using css modules to remove styles pollutioni in component files;
