+++ frontend/src/App.css
@@ -14,15 +14,11 @@
 
 .c-header {
   margin-top: 4rem;
-  color: #ffefdf
+  color: #ffefdf;
 }
 
 .c-qrcode-container {
   margin-top: 2rem;
-}
-
-.c-qrcode {
-  max-width: 100%;
 }
 
 .c-qr-button {
@@ -41,5 +37,5 @@
   height: 1.25rem;
   width: 1.25rem;
   margin-right: 0.5rem;
-  flex: 1
+  flex: 1;
 }
