+++ frontend/src/components/InteractionTemplate/InteractionTemplate.module.css
@@ -1,0 +1,27 @@
+.container {
+  background: #ffefdf;
+  margin: 70px 10px;
+  padding: 30px;
+  box-shadow: 0px 0px 80px 2px gray;
+  border-radius: 40px;
+}
+
+.btn-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  padding: 20px;
+}
+
+.jwt {
+  word-wrap: break-word;
+  max-width: 50vw;
+  white-space: pre-wrap;
+  font-family: monospace;
+}
+
+.qr-code {
+  max-width: 100%;
+}
